% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trails_api.R
\name{get_condtions}
\alias{get_condtions}
\title{Get conditions for specific trails based on trail IDs}
\usage{
get_condtions(project = NULL, key = NULL, ids = NULL)
}
\arguments{
\item{project}{which project you would like to retrieve data from; must be one of the following c("mtb", "hiking", "trailrun", "powder")}

\item{key}{your private key, sign up on adventureproject.com site for one (required)}

\item{ids}{vector of trail IDs (required)}
}
\value{
A data.frame containing the trail conditions. Data.frame will have URL used when calling API in the attributes. Can be retrieved with attr(data, "api_url").
}
\description{
Retrieves a JSON of the currect conditions
}
\examples{
\dontrun{
conditions <- get_condtions(project = project_name, key = key, ids = c("1829020", "1728301")
}

}
