% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountain_api.R
\name{get_routes}
\alias{get_routes}
\title{Get routes for specific Adventure Project based on IDs}
\usage{
get_routes(project = NULL, key = NULL, route_ids = NULL)
}
\arguments{
\item{project}{which project you would like to retrieve data from; must be "mountian".}

\item{key}{your private key, sign up on adventureproject.com site for one (required)}

\item{route_ids}{vector of route IDs (required)}
}
\value{
A data.frame containing all of the trails that match the query parameters.  Data.frame will have URL used when calling API in the attributes. Can be retrieved with attr(data, "api_url").
}
\description{
Retrieves a JSON of routes based on trail IDs.
}
\examples{
\dontrun{
routes <- get_routes(project = "mountian", key = privateKey, route_ids = c("1829020", "1728301"))
}

}
