% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountain_api.R
\name{get_routes_for_lat_lon}
\alias{get_routes_for_lat_lon}
\title{Get the routes for the Mountian Project by longitude and latitude.}
\usage{
get_routes_for_lat_lon(project = NULL, key = NULL, lat = NULL,
  lon = NULL, max_distance = NULL, max_results = NULL,
  min_diff = NULL, max_diff = NULL)
}
\arguments{
\item{project}{which project you would like to retrieve data from; must be "mountian".}

\item{key}{your private key, sign up on adventureproject.com site for one (required)}

\item{lat}{latitude for a given area (required)}

\item{lon}{longitude for given area (required)}

\item{max_distance}{max distance, in miles, from lat, lon. Default: 30. Max: 200.}

\item{max_results}{max number of trails to return. Default: 10. Max: 500.}

\item{min_diff}{min difficulty of routes to return, e.g. 5.6 or V0.}

\item{max_diff}{max difficulty of routes to return, e.g. 5.10a or V2.}
}
\value{
A data.frame containing all of the routes that match the query parameters. data.frame will have URL used when calling API in the attributes. Can be retrieved with attr(data, "api_url").
}
\description{
Retrieves a JSON of routes based on the the longitude and latitude.
}
\examples{
\dontrun{
routes_by_gps <- get_routes_for_lat_lon(project = "mtb", key = privateKey, lat = 31.23242, lon = -118.23421, max_distance = 500, max_results = 175)
}

}
