% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountain_api.R
\name{get_ticks}
\alias{get_ticks}
\title{Returns ticks by user ID or email.}
\usage{
get_ticks(project = NULL, key = NULL, user_id = NULL, email = NULL,
  start_pos = NULL)
}
\arguments{
\item{project}{which project you would like to retrieve data from; must be one of the following c("mtb", "hiking", "trailrun", "powder")}

\item{key}{your private key, sign up on adventureproject.com site for one (required)}

\item{user_id}{the ID of the user to return}

\item{email}{the email address of the user to return}
}
\value{
A data.frame of a users to-dos.
}
\description{
Retrives top 200 ticks for a user. Currently missing the "hardest" and "average" data point.
}
\details{
Either user_id or email is required, but not both.
}
\examples{
\dontrun{
ticks <- get_ticks(project = project_name, key = key, email = "adventure.bro@gmail.com")
}

}
