% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trails_api.R
\name{get_trails_by_id}
\alias{get_trails_by_id}
\title{Get trails for specific Adventure Project based on IDs}
\usage{
get_trails_by_id(project = NULL, key = NULL, ids = NULL)
}
\arguments{
\item{project}{which project you would like to retrieve data from; must be one of the following c("mtb", "hiking", "trailrun", "powder")}

\item{key}{your private key, sign up on adventureproject.com site for one (required)}

\item{ids}{vector of trail IDs (required)}
}
\value{
A data.frame containing all of the trails that match the query parameters.  Data.frame will have URL used when calling API in the attributes. Can be retrieved with attr(data, "api_url").
}
\description{
Retrieves a JSON of trails based on trail IDs.
}
\examples{
\dontrun{
trail_data <- get_trails_by_id(project = "mtb", key = privateKey, ids = c("1829020", "1728301"))
}

}
