% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountain_api.R
\name{get_user}
\alias{get_user}
\title{User information by user ID or email.}
\usage{
get_user(key = NULL, user_id = NULL, email = NULL,
  start_pos = NULL)
}
\arguments{
\item{key}{your private key, sign up on adventureproject.com site for one (required)}

\item{user_id}{the ID of the user to return}

\item{email}{the email address of the user to return}
}
\value{
A data.frame of a users to-dos.
}
\description{
Retrives general user information. Currently only lives under the Mountian Project but returns information about the user from all of hte projects.
}
\details{
Either user_id or email is required, but not both.
}
\examples{
\dontrun{
user <- get_user(project = project_name, key = key, email = "adventure.bro@gmail.com")
}

}
